+++ cmd/cortextool/pkg/client/client.go
@@ -7,6 +7,7 @@ import (
 	"io/ioutil"
 	"net/http"
 	"net/url"
+	"path"
 	"time"
 
 	"github.com/cortexproject/cortex/pkg/util/tls"
@@ -88,7 +89,7 @@ func (r *CortexClient) Query(ctx context.Context, query string) (*http.Response,
 }
 
 func (r *CortexClient) doRequest(path, method string, payload []byte) (*http.Response, error) {
-	req, err := http.NewRequest(method, r.endpoint.String()+path, bytes.NewBuffer(payload))
+	req, err := buildRequest(path, method, r.endpoint, payload)
 	if err != nil {
 		return nil, err
 	}
@@ -153,4 +154,9 @@ func checkResponse(r *http.Response) error {
 	}).Errorln("requests failed")
 
 	return errors.New("failed request to the cortex api")
+}
+
+func buildRequest(p, m string, endpoint *url.URL, payload []byte) (*http.Request, error) {
+	endpoint.Path = path.Join(endpoint.Path, p)
+	return http.NewRequest(m, endpoint.String(), bytes.NewBuffer(payload))
 }